+++ RELEASENOTES.md
@@ -5,7 +5,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 
 ### 0.5.3
 
-* ...
+* Feature - Show app roles in user details
 
 
 ### 0.5.2 (2022-08-23)